+++ project-loader/src/main/kotlin/de/itemis/mps/gradle/logging/IdeaLog.kt
@@ -1,0 +1,10 @@
+package de.itemis.mps.gradle.logging
+
+import com.intellij.openapi.application.PathManager
+import java.io.File
+
+public fun ideaLogFile(): File = File(PathManager.getLogPath(), "idea.log").absoluteFile
+
+public fun printIdeaLogLocation() {
+    System.err.println("MPS log file: ${ideaLogFile().path}")
+}